Retrieve supply, market cap in quotesData

+++ CryptoPortfolio/Networking/Responses/QuotesResponse.swift
@@ -21,9 +21,9 @@ struct QuotesData: Codable {
 	let id: Int
 	let name: String
 	let symbol: String
-	let max_supply: Int?
-	let circulating_supply: Int?
-	let total_supply: Int?
+	let max_supply: Double?
+	let circulating_supply: Double?
+	let total_supply: Double?
 	let cmc_rank: Int
 	let quote: [String: Quote]
 }
